fm後台後端/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
PyMySQL==1.1.2

python-jose[cryptography]
bcrypt
python-multipart

cryptography
pydantic==2.12.5
starlette==0.50.0
anyio==4.12.1
click==8.3.1
h11==0.16.0
idna==3.11
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
annotated-doc==0.0.4
